limpiezas_processor/requirements.txt
google-cloud-storage==2.9.0
